tls_codec: fix: Added bounds check on discriminants
This fixes an UB occuring when any of the following were true:
- In implicit discriminant mode, there are more variants than repr::MAX
- In explicit discriminant mode, the provided value exceeds u32 OR
  repr::MAX
- In path discriminant mode, the resolved value is out of bounds of repr
  (MIN..MAX)

+++ tls_codec/derive/tests/decode.rs
@@ -309,9 +309,10 @@ fn enum_with_data_and_discriminant() {
 mod discriminant {
     pub mod test {
         pub mod constant {
-            pub const TEST_CONST: u8 = 3;
+            pub const TEST_CONST: u16 = 3;
         }
         pub mod enum_val {
+            #[repr(u16)]
             pub enum Test {
                 Potato = 0x0004,
             }
